Bugfix numpy deprecated

+++ muesr/engines/clfc.py
@@ -63,12 +63,12 @@ class LocalFields(object):
             raise ValueError("Must have the same shape!")
         
         
-        self._BLor = np.asarray(BLor,np.float)
-        self._BDip = np.asarray(BDip,np.float)
-        self._BCont = np.asarray(BCont,np.float)
+        self._BLor = np.asarray(BLor,np.float_)
+        self._BDip = np.asarray(BDip,np.float_)
+        self._BCont = np.asarray(BCont,np.float_)
         
         try:
-            self._ACont = np.float(ACont)
+            self._ACont = np.float_(ACont)
         except:
             raise TypeError( "Cannot set value for ACont. Must be float." )
         
@@ -154,7 +154,7 @@ class LocalFields(object):
     @ACont.setter
     def ACont(self,value):
         try:
-            self._ACont = np.float(value)
+            self._ACont = np.float_(value)
         except:
             raise TypeError( "Cannot set value for ACont" )
